Project/requirements.txt
apiritif==0.8.1
Appium-Python-Client==0.40
asn1crypto==0.24.0
astroid==2.2.0
astunparse==1.6.2
atomicwrites==1.3.0
attrs==18.2.0
bzt==1.13.3
certifi==2018.11.29
cffi==1.12.2
chardet==3.0.4
colorama==0.4.1
colorlog==4.0.2
coverage==4.5.2
cryptography==2.6.1
cssselect==1.0.3
decorator==4.4.0
Django==2.1.7
django-cors-headers==2.5.0
django-filter==2.1.0
django-rest-knox==3.6.0
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
EasyProcess==0.2.5
fuzzyset==0.0.19
gevent==1.4.0
greenlet==0.4.15
hdrpy==0.3.3
idna==2.8
ipaddress==1.0.22
isort==4.3.9
jsonpath-rw==1.4.0
lazy-object-proxy==1.3.1
lxml==4.3.2
Markdown==3.0.1
mccabe==0.6.1
more-itertools==6.0.0
mysqlclient==1.4.2.post1
nose==1.3.7
numpy==1.16.2
pluggy==0.9.0
ply==3.11
progressbar33==2.4
psutil==5.6.1
py==1.8.0
pycparser==2.19
PyJWT==1.7.1
pylint==2.3.0
pyOpenSSL==19.0.0
pytest==4.3.0
pytest-cov==2.6.1
pytest-django==3.4.8
python-Levenshtein==0.12.0
pytz==2018.9
PyVirtualDisplay==0.2.1
PyYAML==5.1
requests==2.21.0
selenium==3.141.0
six==1.12.0
SpeechRecognition==3.8.1
terminaltables==3.1.0
texttable==1.6.1
typed-ast==1.3.1
unicodecsv==0.14.1
urllib3==1.24.1
urwid==2.0.1
wrapt==1.11.1
